+++ Assets/ParadoxNotion/NodeCanvas/Tasks/Conditions/Blackboard/CheckFloat.cs
@@ -18,12 +18,14 @@ namespace NodeCanvas.Tasks.Conditions
         [SliderField(0, 0.1f)]
         public float differenceThreshold = 0.05f;
 
-        protected override string info {
+        protected override string info
+        {
             get { return valueA + OperationTools.GetCompareString(checkType) + valueB; }
         }
 
-        protected override bool OnCheck() {
-            return OperationTools.Compare((float)valueA.value, (float)valueB.value, checkType, differenceThreshold);
+        protected override bool OnCheck()
+        {
+            return OperationTools.Compare(valueA.value, valueB.value, checkType, differenceThreshold);
         }
     }
 }
